host: move base64 helpers to fde_key_manager_ca
