+++ host/fde_key_manager/fde_key_manager_ca.h
@@ -18,6 +18,12 @@
 #ifndef FDE_KEY_MANAGER_H_
 #define FDE_KEY_MANAGER_H_
 
+// helper wrapper around mbedtls_base64_encode function
+extern char *basee64_encode(const unsigned char *in_buf, size_t in_buf_len);
+// helper wrapper around mbedtls_base64_decode function
+extern unsigned char *basee64_decode(const char *in_buf,
+                                     size_t in_buf_len, size_t *buf_len);
+
 extern TEEC_Result encrypt_key(unsigned char *in_buf, size_t in_buf_len,
                                unsigned char *handle, size_t *handle_len,
                                unsigned char *out_buf, size_t *out_buf_len);
